Refactoring code

+++ src/main/java/com/vladshkerin/models/Item.java
@@ -72,7 +72,7 @@ public class Item {
         return parentId;
     }
 
-    public void setParentId(int parentId) {
+    public void setParentId(long parentId) {
         this.parentId = parentId;
     }
